Add test for create a parcel with wrong key name  On branch ft-create-a-parcel-4129960  Changes to be committed: 	modified:   app/views/parcels.py 	new file:   tests/test_base.py 	modified:   tests/test_parcels.py
